backend/requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.8.0rc2
distlib==0.3.9
distro==1.9.0
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
django-qrcode==0.3
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.8
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.11.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kombu==5.4.2
Mako==1.3.11
MarkupSafe==3.0.3
mistune==3.2.0
msgpack==1.1.0
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==26.1
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==24.3.0
pyparsing==3.3.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
routing==0.2.0
rpds-py==0.30.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
social-media==0.0.0
SQLAlchemy==2.0.49
sqlparse==0.5.1
tenacity==9.1.4
tqdm==4.67.3
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
websockets==16.0
Werkzeug==3.1.8
WTForms==3.1.2
zope.event==5.0
zope.interface==7.2
